@@ -1,12 +1,11 @@
 import { Request, Response } from "express";
-
+import { container } from "tsyringe";
 import { ListCategoriesUseCase } from "./ListcategoriesUseCase";
 
 class ListCategoriesController {
-  constructor(private ListCategoriesUseCase: ListCategoriesUseCase) {}
-
-  handle(request: Request, response: Response): Response {
-    const all = this.ListCategoriesUseCase.execute();
+  async handle(request: Request, response: Response): Promise<Response> {
+    const listCategoriesUseCase = container.resolve(ListCategoriesUseCase);
+    const all = await listCategoriesUseCase.execute();
 
     return response.json(all);
   }
